@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.3 <3.0.0"
+  sdk: ">=3.1.0 <4.0.0"
 
 
 dependencies:
@@ -16,16 +16,15 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  get: ^4.6.5
-  just_audio: ^0.9.30
-  rxdart: ^0.27.5
-  flutter_media_metadata: ^1.0.0+1
-  path_provider: 2.1.0
-  get_storage: ^2.0.3
-  text_scroll: ^0.1.2
-  permission_handler: ^10.2.0
-
+  cupertino_icons: 1.0.8
+  get: 4.7.2
+  just_audio: 0.10.0
+  rxdart: 0.28.0
+  audio_metadata_reader: 1.4.0
+  path_provider: 2.1.5
+  get_storage: 2.1.1
+  text_scroll: 0.2.0
+  permission_handler: 12.0.0
 
 dev_dependencies:
   flutter_test:
