@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 // this class will hold the configuration files for this app like Theme Data
 
 ThemeData lightTheme = ThemeData(
-  backgroundColor: Colors.grey.shade300, // for timer
+  dialogBackgroundColor: Colors.grey.shade300, // for timer
   brightness: Brightness.light,
   scaffoldBackgroundColor: Colors.grey[300], // for scaffold
   // primaryColor: Colors.grey.shade600,
@@ -46,34 +46,35 @@ ThemeData lightTheme = ThemeData(
     titleLarge: TextStyle(color: Colors.black, fontWeight: FontWeight.bold),
   ),
   checkboxTheme: CheckboxThemeData(
-    fillColor: MaterialStateProperty.resolveWith((states) {
-      if (states.contains(MaterialState.selected)) {
+    fillColor: WidgetStateProperty.resolveWith((states) {
+      if (states.contains(WidgetState.selected)) {
         return Colors.grey.shade600;
       }
       return null;
     }),
   ),
   colorScheme: ColorScheme(
-      secondaryContainer: Colors.black,
-      onPrimaryContainer: Colors.red,
-      primaryContainer: Colors.yellow,
-      brightness: Brightness.light,
-      primary: Colors.grey.shade600, // buttons (textButton's child)
-      onPrimary: Colors.grey.shade400,
-      secondary:
-          Colors.black, // favorite icon, playlist control buttons when enables,
-      onSecondary: Colors.white,
-      error: Colors.red,
-      onError: Colors.white,
-      background: Colors.grey.shade300,
-      onBackground: Colors.grey.shade600,
-      // surface: Colors.grey.shade300,
-      surface: Colors.grey.shade300.withOpacity(0.9),
-      onSurface: Colors.grey.shade600),
+    secondaryContainer: Colors.black,
+    onPrimaryContainer: Colors.red,
+    primaryContainer: Colors.yellow,
+    brightness: Brightness.light,
+    primary: Colors.grey.shade600, // buttons (textButton's child)
+    onPrimary: Colors.grey.shade400,
+    secondary:
+        Colors.black, // favorite icon, playlist control buttons when enables,
+    onSecondary: Colors.white,
+    error: Colors.red,
+    onError: Colors.white,
+    surface: Colors.grey.shade300,
+    onSurface: Colors.grey.shade600,
+    // surface: Colors.grey.shade300,
+    // surface: Colors.grey.shade300.withOpacity(0.9),
+    // onSurface: Colors.grey.shade600,
+  ),
 );
 
 ThemeData darkTheme = ThemeData(
-  backgroundColor: Colors.grey.shade900, // for timer
+  dialogBackgroundColor: Colors.grey.shade900, // for timer
   brightness: Brightness.dark,
   scaffoldBackgroundColor: Colors.grey.shade900, // for scaffold
   // primaryColor: Colors.grey.shade600,
@@ -117,28 +118,29 @@ ThemeData darkTheme = ThemeData(
     titleLarge: TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
   ),
   checkboxTheme: CheckboxThemeData(
-    fillColor: MaterialStateProperty.resolveWith((states) {
-      if (states.contains(MaterialState.selected)) {
+    fillColor: WidgetStateProperty.resolveWith((states) {
+      if (states.contains(WidgetState.selected)) {
         return Colors.grey;
       }
       return null;
     }),
   ),
   colorScheme: ColorScheme(
-      secondaryContainer: Colors.black,
-      onPrimaryContainer: Colors.red,
-      primaryContainer: Colors.yellow,
-      brightness: Brightness.dark,
-      primary: Colors.grey, // buttons (textButton's child)
-      onPrimary: Colors.grey.shade400,
-      secondary:
-          Colors.white, // favorite icon, playlist control buttons when enables,
-      onSecondary: Colors.white,
-      error: Colors.red,
-      onError: Colors.white,
-      background: Colors.grey.shade900,
-      onBackground: Colors.grey.shade600,
-      // surface: Colors.grey.shade300,
-      surface: Colors.grey.shade900.withOpacity(0.9),
-      onSurface: Colors.grey.shade600),
+    secondaryContainer: Colors.black,
+    onPrimaryContainer: Colors.red,
+    primaryContainer: Colors.yellow,
+    brightness: Brightness.dark,
+    primary: Colors.grey, // buttons (textButton's child)
+    onPrimary: Colors.grey.shade400,
+    secondary:
+        Colors.white, // favorite icon, playlist control buttons when enables,
+    onSecondary: Colors.white,
+    error: Colors.red,
+    onError: Colors.white,
+    surface: Colors.grey.shade900,
+    onSurface: Colors.grey.shade600,
+    // surface: Colors.grey.shade300,
+    // surface: Colors.grey.shade900.withOpacity(0.9),
+    // onSurface: Colors.grey.shade600,
+  ),
 );
