@@ -26,7 +26,7 @@ class TimerDisplay extends StatelessWidget {
                 borderRadius: const BorderRadius.only(
                     bottomRight: Radius.circular(12),
                     topRight: Radius.circular(12)),
-                color: Theme.of(context).backgroundColor,
+                color: Theme.of(context).dialogBackgroundColor,
               ),
               child: Padding(
                 padding: const EdgeInsets.only(left: 8.0),
